Updated the congratulations screen per reqs
Canvas says that congratulation message needs to "... display the final score in the center of the screen, at the end of the game. Just use a box that appears in the middle of the screen and have the text show there." So I added the score to the alert message box.